refactor: move operation to it's own module
