+++ rust/lance-table/src/format/transaction/v1/mod.rs
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright The Lance Authors
+
+mod operation;
+
+pub use operation::{validate_operation, Operation, RewriteGroup, RewrittenIndex};
